Update README.md
This commit includes changes to the README file in the repository. The changes include updates to the purpose and description of the repository, as well as the inclusion of information about problem sets, quizzes, and exams. These changes were made in order to provide a more detailed and accurate description of the content and purpose of the repository.